Isolate proc functions into its own runtime module
