+++ opal/runtime/runtime.rb
@@ -2,6 +2,7 @@ require 'runtime/boot'
 require 'runtime/const'
 require 'runtime/send'
 require 'runtime/method'
+require 'runtime/proc'
 require 'runtime/hash'
 require 'runtime/string'
 require 'runtime/regexp'
